Prevent the menu button form unlocking the screen,
When the flag is set in framework there is no way to override the behavior.
Forcing the menu button to always allow the unlock.

Without it set a user property or user data file can be used to re-enable
the use of the menu button.. This can then be coded as a user flag in
settings or other application.