Call deleteIntent when swiping to remove a notification, similar to how the clear button works.
Developers can set a deleteIntent that is called when the clear button is used
to remove all notificaitons, to perform any cleanup needed when deleting
a notificaiton without viewing it.  Since the swipe to remove is basically
a similar use case, but on a single notification, it should call this intent
(if defined) as well.

Change-Id: Idea238d1b8e3b9e5a285d5a8442f193f644f50e5

whitespace errors.